fix(common): flatten tree_sitter module to fix clippy module_inception lint
crates/common/src/tree_sitter/tree_sitter.rs created a module path
(common::tree_sitter::tree_sitter) that clippy's module_inception lint
flags under -D warnings, blocking the pre-push hook. Flatten it into a
single crates/common/src/tree_sitter.rs file and update all
tree_sitter::tree_sitter:: references across the workspace accordingly.
