+++ crates/common/src/lib.rs
@@ -1,4 +1,4 @@
 pub mod analysis_options;
 pub mod model;
-pub mod utils;
 pub mod tree_sitter;
+pub mod utils;
