+++ crates/common/src/model.rs
@@ -1,5 +1,5 @@
 pub mod config_method;
 pub mod diff_aware;
+pub mod language;
 pub mod path_config;
 pub mod position;
-pub mod language;
