+++ crates/static-analysis-kernel/src/analysis.rs
