+++ crates/static-analysis-kernel/src/analysis/ddsa_lib/bridge/query_match.rs
@@ -89,11 +89,11 @@ mod tests {
     use crate::analysis::ddsa_lib::js;
     use crate::analysis::ddsa_lib::test_utils::cfg_test_v8;
     use crate::analysis::ddsa_lib::v8_ds::MirroredVec;
+    use crate::analysis::tree_sitter::{QueryMatch, TSCaptureContent, TSQuery};
     use common::model::language::Language;
-    use common::tree_sitter::tree_sitter::{get_tree};
+    use common::tree_sitter::get_tree;
     use common::utils::position_utils::LineColumnIndex;
     use deno_core::JsRuntime;
-    use crate::analysis::tree_sitter::{QueryMatch, TSCaptureContent, TSQuery};
 
     fn setup_bridge() -> (JsRuntime, QueryMatchBridge, TsNodeBridge) {
         let mut runtime = cfg_test_v8().deno_core_rt();
