+++ crates/static-analysis-kernel/src/analysis/ddsa_lib/context/file_go.rs
@@ -65,7 +65,7 @@ mod tests {
     use crate::analysis::ddsa_lib::context::file_go::FileContextGo;
     use crate::analysis::ddsa_lib::test_utils::cfg_test_v8;
     use common::model::language::Language;
-    use common::tree_sitter::tree_sitter::get_tree;
+    use common::tree_sitter::get_tree;
 
     /// Tests that imports are properly exposed via the context.
     #[test]
