+++ crates/static-analysis-kernel/src/analysis/languages/go/imports.rs
@@ -4,7 +4,7 @@
 
 use crate::analysis::languages::ts_node_text;
 use common::model::language::Language;
-use common::tree_sitter::tree_sitter::{get_tree, get_tree_sitter_language};
+use common::tree_sitter::{get_tree, get_tree_sitter_language};
 use std::sync::LazyLock;
 use tree_sitter::StreamingIterator;
