+++ crates/static-analysis-kernel/src/analysis/languages/python/imports.rs
@@ -4,7 +4,7 @@
 
 use crate::analysis::languages::ts_node_text;
 use common::model::language::Language;
-use common::tree_sitter::tree_sitter::{get_tree, get_tree_sitter_language};
+use common::tree_sitter::{get_tree, get_tree_sitter_language};
 use std::sync::LazyLock;
 use tree_sitter::StreamingIterator;
 
@@ -371,7 +371,7 @@ fn parse_field_child_node<'text>(
 mod tests {
     use super::{parse_imports, parse_imports_with_tree, Entity, Import, ImportEntities};
     use common::model::language::Language;
-    use common::tree_sitter::tree_sitter::get_tree;
+    use common::tree_sitter::get_tree;
 
     /// A shorthand to build an [`Entity`] without an alias.
     pub fn ent(name: &str) -> Entity<'_> {
