+++ crates/static-analysis-kernel/src/analysis/languages/typescript/imports.rs
@@ -4,7 +4,7 @@
 
 use crate::analysis::languages::javascript;
 use common::model::language::Language;
-use common::tree_sitter::tree_sitter::{get_tree, get_tree_sitter_language};
+use common::tree_sitter::{get_tree, get_tree_sitter_language};
 use std::sync::LazyLock;
 
 /// Because the tree-sitter-typescript grammar essentially "inherits" the
