+++ crates/static-analysis-server/src/tree_sitter_tree.rs
@@ -2,7 +2,7 @@ use crate::constants::{ERROR_CODE_NOT_BASE64, ERROR_CODE_NO_ROOT_NODE};
 use crate::model::tree_sitter_tree_node::ServerTreeSitterNode;
 use crate::model::tree_sitter_tree_request::TreeSitterRequest;
 use crate::model::tree_sitter_tree_response::TreeSitterResponse;
-use common::tree_sitter::tree_sitter::{get_tree};
+use common::tree_sitter::get_tree;
 use common::utils::position_utils::LineColumnIndex;
 use kernel::analysis::tree_sitter::map_node;
 use kernel::utils::decode_base64_string;
